+++ src/renderer/src/trayMain.jsx
@@ -1,0 +1,11 @@
+import './index.css'
+
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import Tray from './Tray'
+
+ReactDOM.createRoot(document.getElementById('root')).render(
+  <React.StrictMode>
+    <Tray />
+  </React.StrictMode>
+)